Making dropdown for user

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -29,11 +30,13 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        sleep(2);
+
         $fields = $request->validate([
             'body' => ['required']
         ]);
+        $fields = array_merge($fields, ['user_id'=>Auth::id()]);
 
+        //create the post.
         Post::create($fields);
 
         return redirect('/');
@@ -45,7 +48,13 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return inertia('show', ['post'=>$post]);
+        //eager loading the user relationship to avoid n+1 problem when accessing post's user in the show page
+        //without eager loading, when we access post's user in the show page, it will execute a query to get the user for each post, which will result in n+1 queries (1 query to get the posts and n queries to get the user for each post)
+        $post->load('user');
+
+        $isDeleteAllowed = Auth::user() == $post->user;
+        
+        return inertia('show', ['post'=>$post, 'isDeleteAllowed'=>$isDeleteAllowed]);
     }
 
     /**
@@ -69,6 +78,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        //authroization check to ensure only user who created post can delete it.
+        if (Auth::user() != $post->user) {
+            abort(403);
+        }
+        //delete the post.
         $post->delete();
         return redirect('/');
         dd($post);
